backend service/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
ghostscript==0.7
google-api-core==1.26.3
google-auth==1.28.1
google-cloud-automl==2.2.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-cloud-texttospeech==2.3.0
google-cloud-translate==3.1.0
google-cloud-vision==2.3.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
idna==2.10
libcst==0.3.18
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.25.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
